feat(dropdown): further refine improved select setup
